Use `Option<Weight>` for contract dry-runs
