+++ Cargo.toml
@@ -87,7 +87,6 @@ members = [
 	"frame/collective",
 	"frame/contracts",
 	"frame/contracts/primitives",
-	"frame/contracts/runtime-api",
 	"frame/conviction-voting",
 	"frame/democracy",
 	"frame/fast-unstake",
