--- frame/contracts/runtime-api/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "pallet-contracts-runtime-api"
-version = "4.0.0-dev"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-license = "Apache-2.0"
-homepage = "https://substrate.io"
-repository = "https://github.com/paritytech/substrate/"
-description = "Runtime API definition used to provide dry-run capabilities"
-readme = "README.md"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
-scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
-
-# Substrate Dependencies
-pallet-contracts-primitives = { version = "6.0.0", default-features = false, path = "../primitives" }
-sp-api = { version = "4.0.0-dev", default-features = false, path = "../../../primitives/api" }
-sp-runtime = { version = "6.0.0", default-features = false, path = "../../../primitives/runtime" }
-sp-std = { version = "4.0.0", default-features = false, path = "../../../primitives/std" }
-
-[features]
-default = ["std"]
-std = [
-	"sp-api/std",
-	"codec/std",
-	"scale-info/std",
-	"sp-std/std",
-	"sp-runtime/std",
-	"pallet-contracts-primitives/std",
-]
