--- frame/contracts/runtime-api/README.md
@@ -1,7 +1,0 @@
-Runtime API definition used to provide dry-run capabilities
-
-This API should be imported and implemented by the runtime,
-of a node that wants to provide clients with dry-run
-capabilities.
-
-License: Apache-2.0
