--- frame/contracts/runtime-api/src/lib.rs
@@ -1,85 +1,0 @@
-// This file is part of Substrate.
-
-// Copyright (C) 2019-2022 Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: Apache-2.0
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//! Runtime API definition used to provide dry-run capabilities.
-//!
-//! This API should be imported and implemented by the runtime,
-//! of a node that wants to provide clients with dry-run
-//! capabilities.
-
-#![cfg_attr(not(feature = "std"), no_std)]
-
-use codec::Codec;
-use pallet_contracts_primitives::{
-	Code, CodeUploadResult, ContractExecResult, ContractInstantiateResult, GetStorageResult,
-};
-use sp_std::vec::Vec;
-
-sp_api::decl_runtime_apis! {
-	/// The API to interact with contracts without using executive.
-	pub trait ContractsApi<AccountId, Balance, BlockNumber, Hash> where
-		AccountId: Codec,
-		Balance: Codec,
-		BlockNumber: Codec,
-		Hash: Codec,
-	{
-		/// Perform a call from a specified account to a given contract.
-		///
-		/// See `pallet_contracts::Pallet::call`.
-		fn call(
-			origin: AccountId,
-			dest: AccountId,
-			value: Balance,
-			gas_limit: u64,
-			storage_deposit_limit: Option<Balance>,
-			input_data: Vec<u8>,
-		) -> ContractExecResult<Balance>;
-
-		/// Instantiate a new contract.
-		///
-		/// See `pallet_contracts::Pallet::instantiate`.
-		fn instantiate(
-			origin: AccountId,
-			value: Balance,
-			gas_limit: u64,
-			storage_deposit_limit: Option<Balance>,
-			code: Code<Hash>,
-			data: Vec<u8>,
-			salt: Vec<u8>,
-		) -> ContractInstantiateResult<AccountId, Balance>;
-
-
-		/// Upload new code without instantiating a contract from it.
-		///
-		/// See `pallet_contracts::Pallet::upload_code`.
-		fn upload_code(
-			origin: AccountId,
-			code: Vec<u8>,
-			storage_deposit_limit: Option<Balance>,
-		) -> CodeUploadResult<Hash, Balance>;
-
-		/// Query a given storage key in a given contract.
-		///
-		/// Returns `Ok(Some(Vec<u8>))` if the storage value exists under the given key in the
-		/// specified account and `Ok(None)` if it doesn't. If the account specified by the address
-		/// doesn't exist, or doesn't have a contract then `Err` is returned.
-		fn get_storage(
-			address: AccountId,
-			key: Vec<u8>,
-		) -> GetStorageResult;
-	}
-}
